python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.3
python-dotenv==1.0.0
fastapi-cors==0.0.6
deepface==0.0.79
tensorflow==2.15.0
scikit-learn==1.3.2
matplotlib==3.8.2
hashlib3==2.1.0
aiofiles==23.2.1